+++ gcc/testsuite/g++.dg/contracts/cpp26/assertion-statement.C
@@ -1,0 +1,34 @@
+// generic assertion-statement parsing checks
+// N5008
+// assertion-statement :
+// contract_assert attribute-specifier-seq opt ( conditional-expression ) ;
+// { dg-do compile { target c++23 } }
+// { dg-additional-options "-fcontracts" }
+
+static_assert (__cpp_contracts >= 202502L);
+
+int main()
+{
+  int x;
+
+  contract_assert( x >= 0);
+  contract_assert( x < 0);
+  contract_assert( x == 0);
+
+  contract_assert( x > 0 ? true : false);
+  contract_assert( x < 0 ? true : false);
+
+  contract_assert( x >= 0);
+  contract_assert( x >= 0);
+  contract_assert( x >= 0);
+  contract_assert( x >= 0);
+
+  contract_assert( x >= 0);
+  contract_assert( x < 0);
+  contract_assert( x == 0);
+  contract_assert( x == 1);
+
+
+
+  return 0;
+}
