+++ gcc/testsuite/g++.dg/contracts/cpp26/basic.contract.eval.p6.C
@@ -1,6 +1,10 @@
-//
-// { dg-do compile }
-// { dg-options "-std=c++2a -fcontracts -O -fdump-tree-original -fdump-tree-optimized -Wno-return-type" }
+// N5008 :
+// basic.contract.eval/p6
+// There is an observable checkpoint ([intro.abstract]) C that happens before
+// (evaluation of a contract assertion ) A such that any other operation O
+// that happens before A also happens before C
+// { dg-do compile { target c++23 } }
+// { dg-additional-options "-fcontracts -O -fdump-tree-original -fdump-tree-optimized -Wno-return-type" }
 int f(int& i)
 {
    i++;
