+++ gcc/testsuite/g++.dg/contracts/cpp26/contract-assert-warn-attributes.C
@@ -5,10 +5,3 @@ pre [[unlikley]] ( x > 41 )  // { dg-warning "attributes are ignored on function
 {
   return x + 1;
 }
-
-int main(int ac, char *av[])
-{
-   int i = ac;
-   contract_assert [[deprecated]] (i == 3);  // { dg-warning "attributes are ignored on contract assertions" }
-  return i;
-}
