+++ gcc/testsuite/g++.dg/contracts/cpp26/contracts-friend1.C
@@ -1,6 +1,6 @@
 // ensure contracts on friend declarations are a complete class context
-// { dg-do run }
-// { dg-options "-std=c++2a -fcontracts -fcontract-evaluation-semantic=observe" }
+// { dg-do compile { target c++23 } }
+// { dg-additional-options "-fcontracts -fcontract-evaluation-semantic=observe" }
 
 
 struct X {
