--- gcc/testsuite/g++.dg/contracts/cpp26/empty-nt-param2.C
@@ -1,24 +1,0 @@
-// check that we do not ICE with an empty nontrivial parameter
-// { dg-do compile }
-// { dg-options "-std=c++23 -fcontracts " }
-struct NTClass {
-  NTClass(){};
-  ~NTClass(){};
-};
-
-struct Empty {};
-
-struct PostCond {
-  void f (const NTClass i)
-    post ( true);
-
-  void g (const Empty i)
-    post ( true);
-
-};
-
-void
-PostCond::f (NTClass i){}
-
-void
-PostCond::g (Empty i){}
