+++ gcc/testsuite/g++.dg/contracts/cpp26/lambda.C
@@ -1,5 +1,5 @@
 // check that we do not crash when capturing a constified entity in a contract assertion lambda
-// { dg-do compile }
-// { dg-options "-std=c++2b -fcontracts  " }
+// { dg-do compile { target c++23 } }
+// { dg-additional-options "-fcontracts -fcontract-evaluation-semantic=observe " }
 void f(int i, int j) pre( [i, &j](){ return true;} ( ))
 {}
